--- a/cxotime/__init__.py
+++ b/cxotime/__init__.py
@@ -1,5 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-from .cxotime import CxoTime
+from .cxotime import CxoTime  # noqa
 import ska_helpers
 
 __version__ = ska_helpers.get_version(__package__)