## Description

This effectively does a backport (or sideways port) of https://github.com/astropy/astropy/pull/10360 to speed up parsing of `date` and `greta` format strings by a factor of about 20 and 50, respectively.

For large arrays, the CxoTime object creation time on my Mac is around 300 ns per time.

This also adds a new `maude` format that is basically `greta` as an int64 without the `.`.  It was easiest/fastest to toss this into the parsing PR.

Some timings:

```
In [15]: dates = ['2001:001:01:23:45.666'] * 1_000_000                                                                   

In [16]: from Chandra.Time import date2secs                                                                              

In [17]: time x = date2secs(dates)                                                                                       
CPU times: user 3.49 s, sys: 22.6 ms, total: 3.52 s
Wall time: 3.52 s

In [18]: time x = DateTime(vals, format='date').secs                                                                     
CPU times: user 8.51 s, sys: 20.5 ms, total: 8.53 s
Wall time: 8.53 s

In [19]: time x = CxoTime(vals, format='date').secs                                                                      
CPU times: user 481 ms, sys: 18.4 ms, total: 500 ms
Wall time: 499 ms

In [20]: gretas = ['2001001.012345666'] * 1_000_000                                                                      

In [21]: time x = DateTime(gretas, format='greta').secs                                                                  
CPU times: user 12.8 s, sys: 15.8 ms, total: 12.8 s
Wall time: 12.8 s

In [22]: time x = CxoTime(gretas, format='greta').secs                                                                   
CPU times: user 468 ms, sys: 27 ms, total: 495 ms
Wall time: 495 ms

In [23]: time x = CxoTime(gretas, format='greta').date                                                                   
CPU times: user 4.97 s, sys: 33.6 ms, total: 5 s
Wall time: 5 s

In [24]: time x = DateTime(gretas, format='greta').date                                                                  
CPU times: user 12.9 s, sys: 27.7 ms, total: 12.9 s
Wall time: 12.9 s
```

## Testing

- [x] Passes unit tests on MacOS, Windows
- [x] Functional testing (new tests)