@jeanconn 2020-09-24
Cool!  Though the status of the astropy code isn't clear to me.  Are you sideways-porting to get this in faster here because it will take longer in astropy?  Or because the two types need more special code for handling?  I'm just trying to keep an eye on the things that we might want to pull back *out* of cxotime as you make advances in astropy.time .