@taldcroft 2020-09-25
The astropy 4.2 release may well get dragged out beyond the New Year, and I'd like to get these speed enhancements into CxoTime for shiny. These changes override the base methods in astropy.time so I don't expect any problems, but for sure we'll want to mostly revert this PR when we get 4.2 into Ska3. This PR has been structured with the mixin class to make that process pretty straightforward when we do it.