@arthegost 2020-09-25
Thanks for adding the maude format. It will be helpful in the future when many others are using it in there daily tasks. I wanted to move my comments from the Slack conversations to here.

The maude is returned as a numerical value from the server itself, so it looks like all timestamps will need to be converted before ingested into cxotime.

Also, as mentioned, the timestamps will always come back with the exact same amount of digits, which is 16 (2^4).