@arthegost 2020-09-25
Thank you for the explanation, Tom. It seems that strings and numbers are acceptable for the maude format, as long format='maude' is used.